Vocabulary management: review of additional properties display in HTML. Also, preliminarily including skos:related.
Essentially, I wanted to have narrower and related based on my experience
when mapping Registry subjects to the UAT.  But we can't have potentially
many links to other terms in "More" in the default table, so I'm now grouping
the links by property, and where there are objects, you can pop the links up.

Let's see if people can stand it.