Now adding useInstead into desise.
It has been an oversight that it is missing, and adding it is not against
the spec, so let's just do it.

Somewhat more fishily, this fixes an obvious typo in the useInstead
of refframe's #supergalactic.  Theoretically, this would take a VEP.
But then: there's no way this could break anything at this point, and
it's obvious that it needs to be done.

In the future, there's a validator...