Writing RDF/X with rdflib rather than rapper now.
This removes a non-python dependency, and an exec call on top.

The literal form of the RDF/X is different from what we had before, but
eyeballing the RDF confirms the expectation that the triples are equivalent,
if somewhat harder to pull out by non-RDF tooling.  But for that we're
offering desise, and Vocabularies 2 doesn't make any guarantees for our RDF/X
beyond "it's valid RDF/X for the vocabulary triples".