We now have copyable concept URIs under the terms.
People too often copy the URI of a term on one of ourHTML pages, and it
should be a bit easier to do the right thing, i.e., copy the unversioned URI.

Whether this helps *a lot* I can't say.  I've tried putting in the full URI
in a separate, colpanning row, but that sucks with :target, and I don't
see a solution of that.  One could consider playing games with mouse-following
URIs, but then for *us* the full URIs rarely matter, so they sholdn't
grab up too much attention and/or space.

Ah well.