Adding support for skos:topconcept.
This is done through a new key in vocabs.conf; the alternative would
have been some trick to allow extra relationships in terms.csv with the
current term as object; but I think it's really preferable to add extra
"global" properties in vocabs.conf.

This will currently only work with SKOS vocabularies because the others
don't define the skos: prefix.  This would be simple to fix, but
currently I'm leaning towards rejecting topconcept for non-skos vocabularies.