Now adding onclick handlers to concept links.
This is to avoid going through redirects when people click on them.
Ever since we give URLs as hrefs for the terms ("copy for the concept URI"),
when someone clicked on them, there was a server roundtrip with all the
redirects (because the URI the browser sees is the versioned URI, not
the vocabulary URI).  There is little we can do about this in general
(which is slightly regrettable because at least webkit browsers sometimes
get confused with the fragments when they get the http->https upgrade
redirects).  But when people have javascript on, we can avoid this;
this is what this commit does.